added conditioned discount to gui

+++ src/ServiceLayer/StoreRoleImp.java
@@ -23,7 +23,7 @@ public class StoreRoleImp implements IStoreRole {
 
 
     @Override
-    public boolean add_store_product(String user_name, String store_name, String product_name, int product_price, int product_amount) {
+    public boolean add_store_product(String user_name, String store_name, String product_name, double product_price, int product_amount) {
         my_log.logger.info("add_store_product");
         if (!SystemManage_Facade.is_initialized() || product_amount == 0) return false;
         if (SystemManage_Facade.find_subscriber(user_name) && SubscribersManage_Facade.check_if_logged_in(user_name)) {
